+++ test/unit/integrations/sequel_test.rb
@@ -28,12 +28,15 @@ module SequelTest
           primary_key :id
           column :state, :string
         end if create_table
-        model = Class.new(Sequel::Model(DB[table_identifier])) do
+        model = Class.new(Sequel::Model) do
           self.raise_on_save_failure = false
           (class << self; self; end).class_eval do
             define_method(:name) { "SequelTest::#{table_name.to_s.capitalize}" }
           end
+          
+          set_dataset(DB[table_identifier])
         end
+        model.set_dataset(DB[table_identifier])
         model.class_eval(&block) if block_given?
         model
       end
@@ -321,18 +324,29 @@ module SequelTest
   
   class MachineWithColumnDefaultTest < BaseTestCase
     def setup
-      @model = new_model
-      DB.alter_table :foo do
-        add_column :status, :string, :default => 'idling'
+      @original_stderr, $stderr = $stderr, StringIO.new
+      
+      DB.create_table!(::Sequel::SQL::Identifier.new(:foo)) do
+        primary_key :id
+        column :state, :string, :default => 'idling'
       end
+      @model = new_model(false)
       @model.class_eval { get_db_schema(true) }
       
-      @machine = StateMachine::Machine.new(@model, :status, :initial => :parked)
+      @machine = StateMachine::Machine.new(@model, :initial => :parked)
       @record = @model.new
     end
     
     def test_should_use_machine_default
-      assert_equal 'parked', @record.status
+      assert_equal 'parked', @record.state
+    end
+    
+    def test_should_generate_a_warning
+      assert_match(/Both SequelTest::Foo and its :state machine have defined a default for "state". Use only one or the other for defining defaults to avoid unexpected behaviors\./, $stderr.string)
+    end
+    
+    def teardown
+      $stderr = @original_stderr
     end
   end
   
@@ -495,7 +509,7 @@ module SequelTest
     def setup
       @model = new_model
       DB.alter_table :foo do
-        add_column :status, :string, :default => 'idling'
+        add_column :status, :string
       end
       @model.class_eval { get_db_schema(true) }
       
@@ -610,7 +624,7 @@ module SequelTest
     def setup
       @model = new_model
       DB.alter_table :foo do
-        add_column :status, :string, :default => 'idling'
+        add_column :status, :string
       end
       @model.class_eval { get_db_schema(true) }
       
@@ -633,7 +647,7 @@ module SequelTest
     def setup
       @model = new_model
       DB.alter_table :foo do
-        add_column :status, :string, :default => 'idling'
+        add_column :status, :string
       end
       @model.class_eval { get_db_schema(true) }